MVP
This is a basic implementation of the game of life in Processing3 using
the Python Mode.

There is a custom class called Cell containing the rules of the game.
Each instance of Cell is either alive or dead and contains its
coordinates on a 60x60 toroidal grid.

Custom functions are defined in the processing sketch to update the
board on each draw() loop.