Refactoring
Abstracted the methods previously written in the sketch to a new class
named Board.

An instance of Board keeps track of the coordinates of the cell and is
also responsible for their update and show methods.

This halves the lines of the sketch file and makes will make it easier
to add interactivity to the programme.