Tracker Cells
First truly new cell and board class.
Right click on a TrackerBoard cell and it turns into a tracker: it is
of a different colour, and so is any cell born in its neighbourhood.
If the tracker does not immediately die, there is a good chance that
its progeny will reach all of the board!